Changed RenderManager.cs to have its ClearScreen method as public.  Also removed the ClearScreen call from the DrawScene method.
Changed Game.cs to clear the screen with the RenderManager's ClearScreen method *BEFORE* drawing any of the other required game components.

These changes are necessary because we were drawing the 2d component, clearing the screen, and then drawing the 3d component.  This was causing the 2d component to get overwritten.  Putting the ClearScreen call in the Game class seems to be the only way to avoid this problem currently.